+++ flash/core/data/input_transform.py
@@ -14,7 +14,7 @@
 import inspect
 from dataclasses import dataclass
 from functools import partial, wraps
-from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Type, Union
+from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Union
 
 from pytorch_lightning.utilities.enums import LightningEnum
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
@@ -23,17 +23,20 @@ from torch.utils.data._utils.collate import default_collate
 import flash
 from flash.core.data.callback import FlashCallback
 from flash.core.data.io.input import DataKeys
-from flash.core.data.io.input_transform import _InputTransformProcessor
-from flash.core.data.properties import Properties
-from flash.core.data.states import CollateFn
+from flash.core.data.io.input_transform import _InputTransformProcessorV2
+from flash.core.data.properties import ProcessState, Properties
+from flash.core.data.states import (
+    CollateFn,
+    PerBatchTransform,
+    PerBatchTransformOnDevice,
+    PerSampleTransform,
+    PerSampleTransformOnDevice,
+)
 from flash.core.data.transforms import ApplyToKeys
 from flash.core.data.utils import _INPUT_TRANSFORM_FUNCS, _STAGES_PREFIX
 from flash.core.registry import FlashRegistry
 from flash.core.utilities.stages import RunningStage
-
-INPUT_TRANSFORM_TYPE = Optional[
-    Union[Type["InputTransform"], Callable, Tuple[Union[LightningEnum, str], Dict[str, Any]], Union[LightningEnum, str]]
-]
+from flash.core.utilities.types import INPUT_TRANSFORM_TYPE
 
 
 class InputTransformPlacement(LightningEnum):
@@ -756,9 +759,10 @@ class InputTransform(Properties):
 
     @partial(transform_context, current_fn="per_sample_transform")
     def _per_sample_transform(self, sample: Any) -> Any:
+        fn = self._get_current_transform(PerSampleTransform)
         if isinstance(sample, list):
-            return [self.current_transform(s) for s in sample]
-        return self.current_transform(sample)
+            return [fn(s) for s in sample]
+        return fn(sample)
 
     @partial(transform_context, current_fn="per_batch_transform")
     def _per_batch_transform(self, batch: Any) -> Any:
@@ -767,7 +771,7 @@ class InputTransform(Properties):
         .. note::     This option is mutually exclusive with :meth:`per_sample_transform_on_device`,     since if both
         are specified, uncollation has to be applied.
         """
-        return self.current_transform(batch)
+        return self._get_current_transform(PerBatchTransform)(batch)
 
     @partial(transform_context, current_fn="collate")
     def _collate(self, samples: Sequence, metadata=None) -> Any:
@@ -789,6 +793,12 @@ class InputTransform(Properties):
             return collate_fn(samples, metadata)
         return collate_fn(samples)
 
+    def _get_current_transform(self, process_state: ProcessState):
+        fn = self.get_state(process_state)
+        if fn is not None and fn.transform is not None:
+            return fn.transform
+        return self.current_transform
+
     @partial(transform_context, current_fn="per_sample_transform_on_device")
     def _per_sample_transform_on_device(self, sample: Any) -> Any:
         """Transforms to apply to the data before the collation (per-sample basis).
@@ -798,9 +808,10 @@ class InputTransform(Properties):
         workers, since to make that happen     each of the workers would have to create it's own CUDA-context which
         would pollute GPU memory (if on GPU).
         """
+        fn = self._get_current_transform(PerSampleTransformOnDevice)
         if isinstance(sample, list):
-            return [self.current_transform(s) for s in sample]
-        return self.current_transform(sample)
+            return [fn(s) for s in sample]
+        return fn(sample)
 
     @partial(transform_context, current_fn="per_batch_transform_on_device")
     def _per_batch_transform_on_device(self, batch: Any) -> Any:
@@ -809,7 +820,7 @@ class InputTransform(Properties):
         .. note::     This function won't be called within the dataloader workers, since to make that happen     each of
         the workers would have to create it's own CUDA-context which would pollute GPU memory (if on GPU).
         """
-        return self.current_transform(batch)
+        return self._get_current_transform(PerBatchTransformOnDevice)(batch)
 
     #############
     # UTILITIES #
@@ -976,18 +987,25 @@ def create_transform(
     running_stage: RunningStage,
     data_pipeline_state: Optional["flash.core.data.data_pipeline.DataPipelineState"] = None,
     input_transforms_registry: Optional[FlashRegistry] = None,
+    transform_kwargs: Optional[Dict] = None,
 ) -> Optional["InputTransform"]:
+
+    if not transform_kwargs:
+        transform_kwargs = {}
 
     if isinstance(transform, InputTransform):
         transform._data_pipeline_state = data_pipeline_state
         return transform
 
     if inspect.isclass(transform) and issubclass(transform, InputTransform):
-        return transform(running_stage=running_stage, data_pipeline_state=data_pipeline_state)
+        return transform(running_stage=running_stage, data_pipeline_state=data_pipeline_state, **transform_kwargs)
 
     if isinstance(transform, Callable):
         return LambdaInputTransform(
-            running_stage=running_stage, transform=transform, data_pipeline_state=data_pipeline_state
+            running_stage=running_stage,
+            transform=transform,
+            data_pipeline_state=data_pipeline_state,
+            **transform_kwargs,
         )
 
     if isinstance(transform, tuple) or isinstance(transform, (LightningEnum, str)):
@@ -1009,9 +1027,9 @@ def _make_collates(input_transform: "InputTransform", on_device: bool, collate: 
 
 def _create_collate_input_transform_processors(
     input_transform: "InputTransform", callbacks: List[FlashCallback]
-) -> Tuple[_InputTransformProcessor, _InputTransformProcessor]:
-    """This utility is used to create the 2 `_InputTransformProcessor` objects which contain the transforms used as
-    the DataLoader `collate_fn` and the DataModule `on_after_batch_transfer` hook."""
+) -> Tuple[_InputTransformProcessorV2, _InputTransformProcessorV2]:
+    """This utility is used to create the 2 `_InputTransformProcessorV2` objects which contain the transforms used
+    as the DataLoader `collate_fn` and the DataModule `on_after_batch_transfer` hook."""
 
     from flash.core.data.data_pipeline import DataPipeline
 
@@ -1042,18 +1060,18 @@ def _create_collate_input_transform_processors(
         )
 
     worker_collate_fn = (
-        worker_collate_fn.collate_fn if isinstance(worker_collate_fn, _InputTransformProcessor) else worker_collate_fn
+        worker_collate_fn.collate_fn if isinstance(worker_collate_fn, _InputTransformProcessorV2) else worker_collate_fn
     )
 
-    worker_input_transform_processor = _InputTransformProcessor(
+    worker_input_transform_processor = _InputTransformProcessorV2(
         input_transform,
         worker_collate_fn,
-        input_transform._per_sample_transform,
+        input_transform._identity if input_transform.serving else input_transform._per_sample_transform,
         input_transform._per_batch_transform,
         input_transform.running_stage,
         callbacks=callbacks,
     )
-    device_input_transform_processor = _InputTransformProcessor(
+    device_input_transform_processor = _InputTransformProcessorV2(
         input_transform,
         device_collate_fn,
         input_transform._per_sample_transform_on_device,