+++ flash_examples/image_classification.py
@@ -23,6 +23,8 @@ download_data("https://pl-flash-data.s3.amazonaws.com/hymenoptera_data.zip", "./
 datamodule = ImageClassificationData.from_folders(
     train_folder="data/hymenoptera_data/train/",
     val_folder="data/hymenoptera_data/val/",
+    batch_size=4,
+    transform_kwargs={"image_size": (196, 196)},
 )
 
 # 2. Build the task
@@ -38,7 +40,8 @@ datamodule = ImageClassificationData.from_files(
         "data/hymenoptera_data/val/bees/65038344_52a45d090d.jpg",
         "data/hymenoptera_data/val/bees/590318879_68cf112861.jpg",
         "data/hymenoptera_data/val/ants/540543309_ddbb193ee5.jpg",
-    ]
+    ],
+    batch_size=3,
 )
 predictions = trainer.predict(model, datamodule=datamodule)
 print(predictions)