+++ flash_examples/integrations/baal/image_classification_active_learning.py
@@ -25,8 +25,7 @@ download_data("https://pl-flash-data.s3.amazonaws.com/hymenoptera_data.zip", "./
 # Implement the research use-case where we mask labels from labelled dataset.
 datamodule = ActiveLearningDataModule(
     ImageClassificationData.from_folders(train_folder="data/hymenoptera_data/train/", batch_size=2),
-    initial_num_labels=5,
-    val_split=0.1,
+    initial_num_labels=10,
 )
 
 # 2. Build the task
@@ -51,7 +50,8 @@ trainer.finetune(model, datamodule=datamodule, strategy="freeze")
 
 # 4. Predict what's on a few images! ants or bees?
 datamodule = ImageClassificationData.from_files(
-    predict_files=["data/hymenoptera_data/val/bees/65038344_52a45d090d.jpg"]
+    predict_files=["data/hymenoptera_data/val/bees/65038344_52a45d090d.jpg"],
+    batch_size=1,
 )
 predictions = trainer.predict(model, datamodule=datamodule)
 print(predictions)