+++ tests/image/classification/test_data.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import csv
+from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, List, Tuple
 
@@ -21,21 +22,19 @@ import torch
 import torch.nn as nn
 
 from flash.core.data.io.input import DataKeys
-from flash.core.data.transforms import ApplyToKeys, merge_transforms
+from flash.core.data.transforms import ApplyToKeys
 from flash.core.utilities.imports import (
-    _ALBUMENTATIONS_AVAILABLE,
     _FIFTYONE_AVAILABLE,
     _IMAGE_AVAILABLE,
     _MATPLOTLIB_AVAILABLE,
     _PIL_AVAILABLE,
     _TORCHVISION_AVAILABLE,
 )
-from flash.image import ImageClassificationData
-from flash.image.classification.transforms import AlbumentationsAdapter, default_transforms
+from flash.image import ImageClassificationData, ImageClassificationInputTransform
 from tests.helpers.utils import _IMAGE_TESTING
 
 if _TORCHVISION_AVAILABLE:
-    import torchvision
+    import torchvision.transforms as T
     from torchvision.datasets import FakeData
 
 if _PIL_AVAILABLE:
@@ -43,9 +42,6 @@ if _PIL_AVAILABLE:
 
 if _FIFTYONE_AVAILABLE:
     import fiftyone as fo
-
-if _ALBUMENTATIONS_AVAILABLE:
-    import albumentations
 
 
 def _dummy_image_loader(_):
@@ -193,7 +189,7 @@ def test_from_filepaths_visualise_multilabel(tmpdir):
         test_files=[image_b, image_b],
         test_targets=[[0, 0, 1], [1, 1, 0]],
         batch_size=2,
-        image_size=(64, 64),
+        transform_kwargs={"image_size": (64, 64)},
     )
     # disable visualisation for testing
     assert dm.data_fetcher.block_viz_window is True
@@ -225,12 +221,10 @@ def test_from_filepaths_splits(tmpdir):
 
     assert len(train_filepaths) == len(train_labels)
 
-    _to_tensor = {
-        "per_sample_transform": nn.Sequential(
-            ApplyToKeys(DataKeys.INPUT, torchvision.transforms.ToTensor()),
-            ApplyToKeys(DataKeys.TARGET, torch.as_tensor),
-        ),
-    }
+    _to_tensor = nn.Sequential(
+        ApplyToKeys(DataKeys.INPUT, T.Compose([T.ToTensor(), T.Resize(img_size)])),
+        ApplyToKeys(DataKeys.TARGET, torch.as_tensor),
+    )
 
     def run(transform: Any = None):
         dm = ImageClassificationData.from_files(
@@ -241,7 +235,6 @@ def test_from_filepaths_splits(tmpdir):
             batch_size=B,
             num_workers=0,
             val_split=val_split,
-            image_size=img_size,
         )
         data = next(iter(dm.train_dataloader()))
         imgs, labels = data["input"], data["target"]
@@ -266,10 +259,10 @@ def test_from_folders_only_train(tmpdir):
 
     img_data = ImageClassificationData.from_folders(train_dir, train_transform=None, batch_size=1)
 
-    data = next(iter(img_data.train_dataloader()))
+    data = img_data.train_dataset[0]
     imgs, labels = data["input"], data["target"]
-    assert imgs.shape == (1, 3, 196, 196)
-    assert labels.shape == (1,)
+    assert isinstance(imgs, Image.Image)
+    assert labels == 0
 
 
 @pytest.mark.skipif(not _IMAGE_TESTING, reason="image libraries aren't installed.")
@@ -454,6 +447,7 @@ def test_from_fiftyone(tmpdir):
 
 @pytest.mark.skipif(not _IMAGE_TESTING, reason="image libraries aren't installed.")
 def test_from_datasets():
+
     img_data = ImageClassificationData.from_datasets(
         train_dataset=FakeData(size=3, num_classes=2),
         val_dataset=FakeData(size=3, num_classes=2),
@@ -569,28 +563,27 @@ def test_from_bad_csv_no_image(bad_csv_no_image):
 
 
 @pytest.mark.skipif(not _IMAGE_AVAILABLE, reason="image libraries aren't installed.")
-@pytest.mark.skipif(not _ALBUMENTATIONS_AVAILABLE, reason="albumentations isn't installed.")
-def test_albumentations_mixup(single_target_csv):
-    def mixup(batch, alpha=1.0):
-        images = batch["input"]
-        targets = batch["target"].float().unsqueeze(1)
+def test_mixup(single_target_csv):
+    @dataclass
+    class MyTransform(ImageClassificationInputTransform):
 
-        lam = np.random.beta(alpha, alpha)
-        perm = torch.randperm(images.size(0))
+        alpha: float = 1.0
 
-        batch["input"] = images * lam + images[perm] * (1 - lam)
-        batch["target"] = targets * lam + targets[perm] * (1 - lam)
-        for e in batch["metadata"]:
-            e.update({"lam": lam})
-        return batch
+        def mixup(self, batch):
+            images = batch["input"]
+            targets = batch["target"].float().unsqueeze(1)
 
-    train_transform = {
-        # applied only on images as ApplyToKeys is used with `input`
-        "per_sample_transform": ApplyToKeys("input", AlbumentationsAdapter(albumentations.HorizontalFlip(p=0.5))),
-        "per_batch_transform": mixup,
-    }
-    # merge the default transform for this task with new one.
-    train_transform = merge_transforms(default_transforms((256, 256)), train_transform)
+            lam = np.random.beta(self.alpha, self.alpha)
+            perm = torch.randperm(images.size(0))
+
+            batch["input"] = images * lam + images[perm] * (1 - lam)
+            batch["target"] = targets * lam + targets[perm] * (1 - lam)
+            for e in batch["metadata"]:
+                e.update({"lam": lam})
+            return batch
+
+        def per_batch_transform(self):
+            return self.mixup
 
     img_data = ImageClassificationData.from_csv(
         "image",
@@ -598,7 +591,7 @@ def test_albumentations_mixup(single_target_csv):
         train_file=single_target_csv,
         batch_size=2,
         num_workers=0,
-        train_transform=train_transform,
+        train_transform=MyTransform,
     )
 
     batch = next(iter(img_data.train_dataloader()))